+++ Sources/SomnioProtocol/Payloads/Position.swift
@@ -7,10 +7,12 @@ public struct PositionMessage: Codable, Sendable, Equatable {
     public var entityIndex: Int16
     public var x: Int16
     public var y: Int16
-    public var facing: Int16
+    /// Continuous heading in degrees `[0, 360)` (0° = south, 90° = east); the receiver
+    /// normalizes via `Heading(degrees:)`.
+    public var facing: Float
     public var tempo: Int16
 
-    public init(entityIndex: Int16, x: Int16, y: Int16, facing: Int16, tempo: Int16) {
+    public init(entityIndex: Int16, x: Int16, y: Int16, facing: Float, tempo: Int16) {
         self.entityIndex = entityIndex
         self.x = x
         self.y = y
